Use FromStr instead of FromOsStr
FromOsStr was really designed to deal with OsString specific types
and I'm planning to release it in the upcoming version. If you don't
need to consume underlying OsString you an use FromUtf8 alias like so

Code should be a bit more readable plus you will be able to share
impl FromStr for OutputStyle with other parts of your code.

Alternatively you can consider to derive Bpaf for OutputStyle  and use
```rust
    #[bpaf(extern(output_style), fallback(OutputStyle::Bordered))]
    pub style: OutputStyle,
```

This should give  program 3 flags: --simple, --multiline and --bordered

But that's a matter of preference